Refactor: create a module for the league_stat class helper methods
